chore: move files to correct package
Move Jobviewmodel and job compose screen to respective correct package structure
