+++ app/src/main/java/seek/codingtask/searchresults/presentation/SearchResultsNavAction.kt
@@ -3,7 +3,7 @@ package seek.codingtask.searchresults.presentation
 import androidx.navigation.NavController
 import com.seek.android.core.presentation.mvi.NavigationAction
 import seek.codingtask.jobdetails.JobDetailsArgs
-import seek.codingtask.jobdetails.JobDetailsScreen
+import seek.codingtask.jobdetails.presentation.views.compose.JobDetailsScreen
 import timber.log.Timber
 
 sealed class SearchResultsNavAction : NavigationAction {
